Changed code with respect to the review
